+++ library/trans.c
@@ -255,7 +255,8 @@ static int dnet_trans_resend(struct dnet_trans *t)
 	t->st = st;
 
 	dnet_log(n, DNET_LOG_INFO, "%s: resending transaction %llu -> %s.\n",
-			dnet_dump_id(t->cmd.id), t->trans, dnet_state_dump_addr(st));
+			dnet_dump_id(t->cmd.id), (unsigned long long)t->trans,
+			dnet_state_dump_addr(st));
 
 	dnet_data_ready(t->st, &t->r);
 	return 1;
